[libs] E: Moving pthread to syscall
This commit moves the pthread module from the posix crate to the syscall
crate.
