+++ src/libs/posix/src/pthread/mutex.rs
@@ -82,140 +82,6 @@ pub unsafe extern "C" fn pthread_mutexattr_destroy(attr: *mut pthread_mutexattr_
 }
 
 //==================================================================================================
-// pthread_mutex_destroy()
-//==================================================================================================
-
-///
-/// # Description
-///
-/// Destroys a mutex.
-///
-/// # Parameters
-///
-/// - `mutex`: Mutex object.
-///
-/// # Returns
-///
-/// If successful, zero is returned. Otherwise, an error code is returned instead.
-///
-/// # Safety
-///
-/// This function is unsafe because it may dereference raw pointers.
-///
-/// It is safe to call this function if the following conditions are met:
-///
-/// - `mutex` points to a valid `pthread_mutex_t` structure.
-///
-#[unsafe(no_mangle)]
-pub unsafe extern "C" fn pthread_mutex_destroy(mutex: *mut pthread_mutex_t) -> c_int {
-    // Check if `mutex` is not valid.
-    if mutex.is_null() {
-        ::syslog::error!("pthread_mutex_destroy(): invalid mutex pointer");
-        return ErrorCode::InvalidArgument.get();
-    }
-
-    if let Err(error) = pthread::pthread_mutex_destroy(&mut *mutex) {
-        return error.code.get();
-    }
-
-    0
-}
-
-//==================================================================================================
-// pthread_mutex_init()
-//==================================================================================================
-
-///
-/// # Description
-///
-/// Initializes a mutex.
-///
-/// # Parameters
-///
-/// - `mutex`: Mutex object.
-/// - `attr`: Mutex attributes.
-///
-/// # Returns
-///
-/// If successful, zero is returned. Otherwise, an error code is returned instead.
-///
-/// # Safety
-///
-/// This function is unsafe because it may dereference raw pointers.
-///
-/// It is safe to call this function if the following conditions are met:
-///
-/// - `mutex` points to a valid `pthread_mutex_t` structure.
-///
-#[unsafe(no_mangle)]
-pub unsafe extern "C" fn pthread_mutex_init(
-    mutex: *mut pthread_mutex_t,
-    attr: *const pthread_mutexattr_t,
-) -> c_int {
-    // Check if `mutex` is not valid.
-    if mutex.is_null() {
-        ::syslog::error!("pthread_mutex_init(): invalid mutex pointer");
-        return ErrorCode::InvalidArgument.get();
-    }
-
-    // Check if we should use custom attributes.
-    if !attr.is_null() {
-        ::syslog::warn!("pthread_mutex_init(): custom attributes not supported, ignoring");
-    }
-
-    // TODO: once we support custom attributes, dereference that pointer.
-    let attr: pthread_mutexattr_t = pthread_mutexattr_t::default();
-
-    if let Err(error) = pthread::pthread_mutex_init(&mut *mutex, &attr) {
-        return error.code.get();
-    }
-
-    0
-}
-
-//==================================================================================================
-// pthread_mutex_lock()
-//==================================================================================================
-
-///
-/// # Description
-///
-/// Locks a mutex.
-///
-/// # Parameters
-///
-/// - `mutex`: Mutex object.
-///
-/// # Returns
-///
-/// If successful, zero is returned. Otherwise, an error code is returned instead.
-///
-/// # Safety
-///
-/// This function is unsafe because it may dereference raw pointers.
-///
-/// It is safe to call this function if the following conditions are met:
-///
-/// - `mutex` points to a valid `pthread_mutex_t` structure.
-///
-#[unsafe(no_mangle)]
-pub unsafe extern "C" fn pthread_mutex_lock(mutex: *mut pthread_mutex_t) -> c_int {
-    // Check if `mutex` is not valid.
-    if mutex.is_null() {
-        ::syslog::error!("pthread_mutex_lock(): invalid mutex pointer");
-        return ErrorCode::InvalidArgument.get();
-    }
-
-    match pthread::pthread_mutex_lock(&mut *mutex) {
-        Ok(_) => 0,
-        Err(error) => {
-            ::syslog::error!("pthread_mutex_lock(): failed to lock mutex (error={:?})", error);
-            error.code.get()
-        },
-    }
-}
-
-//==================================================================================================
 // pthread_mutex_timedlock()
 //==================================================================================================
 
@@ -323,44 +189,5 @@ pub unsafe extern "C" fn pthread_mutex_trylock(mutex: *mut pthread_mutex_t) -> c
             ::syslog::error!("pthread_mutex_trylock(): failed to lock mutex (error={:?})", error);
             error.code.get()
         },
-    }
-}
-
-//==================================================================================================
-// pthread_mutex_unlock()
-//==================================================================================================
-
-///
-/// # Description
-///
-/// Unlocks a mutex.
-///
-/// # Parameters
-///
-/// - `mutex`: Mutex object.
-///
-/// # Returns
-///
-/// If successful, zero is returned. Otherwise, an error code is returned instead.
-///
-/// # Safety
-///
-/// This function is unsafe because it may dereference raw pointers.
-///
-/// It is safe to call this function if the following conditions are met:
-///
-/// - `mutex` points to a valid `pthread_mutex_t` structure.
-///
-#[unsafe(no_mangle)]
-pub unsafe extern "C" fn pthread_mutex_unlock(mutex: *mut pthread_mutex_t) -> c_int {
-    // Check if `mutex` is not valid.
-    if mutex.is_null() {
-        ::syslog::error!("pthread_mutex_unlock(): invalid mutex pointer");
-        return ErrorCode::InvalidArgument.get();
-    }
-
-    match pthread::pthread_mutex_unlock(&mut *mutex) {
-        Ok(_) => 0,
-        Err(error) => error.code.get(),
     }
 }
