+++ src/libs/syscall/src/pthread/bindings/mod.rs
@@ -1,0 +1,36 @@
+// Copyright(c) The Maintainers of Nanvix.
+// Licensed under the MIT License.
+
+//==================================================================================================
+// Modules
+//==================================================================================================
+
+pub mod pthread_attr_destroy;
+pub mod pthread_attr_init;
+pub mod pthread_attr_setstacksize;
+pub mod pthread_cond_broadcast;
+pub mod pthread_cond_destroy;
+pub mod pthread_cond_init;
+pub mod pthread_cond_signal;
+pub mod pthread_cond_timedwait;
+pub mod pthread_cond_wait;
+pub mod pthread_create;
+pub mod pthread_getschedparam;
+pub mod pthread_getspecific;
+pub mod pthread_join;
+pub mod pthread_key_create;
+pub mod pthread_kill;
+pub mod pthread_mutex_destroy;
+pub mod pthread_mutex_init;
+pub mod pthread_mutex_lock;
+pub mod pthread_mutex_unlock;
+pub mod pthread_rwlock_rdlock;
+pub mod pthread_rwlock_wrlock;
+pub mod pthread_self;
+pub mod pthread_setcancelstate;
+pub mod pthread_setspecific;
+pub mod pthread_sigmask;
+pub mod sem_destroy;
+pub mod sem_init;
+pub mod sem_post;
+pub mod sem_wait;
