+++ src/libs/syscall/src/pthread/bindings/pthread_kill.rs
@@ -7,19 +7,18 @@
 
 use ::sysapi::{
     ffi::c_int,
-    sys_types::pthread_key_t,
+    sys_types::pthread_t,
 };
-use ::syscall::pthread;
 
 //==================================================================================================
-// pthread_key_delete()
+// Standalone Functions
 //==================================================================================================
 
+// TODO: add description
 #[allow(clippy::missing_safety_doc)]
 #[unsafe(no_mangle)]
-pub unsafe extern "C" fn pthread_key_delete(key: pthread_key_t) -> c_int {
-    match pthread::pthread_key_delete(key) {
-        Ok(()) => 0,
-        Err(error) => error.code.get(),
-    }
+pub unsafe extern "C" fn pthread_kill(_thread: pthread_t, _sig: c_int) -> c_int {
+    // TODO: https://github.com/nanvix/nanvix/issues/716
+    ::syslog::warn!("pthread_kill(): not implemented");
+    0
 }
