+++ src/libs/syscall/src/pthread/bindings/pthread_mutex_lock.rs
@@ -1,0 +1,54 @@
+// Copyright(c) The Maintainers of Nanvix.
+// Licensed under the MIT License.
+
+//==================================================================================================
+// Imports
+//==================================================================================================
+
+use ::sys::error::ErrorCode;
+use ::sysapi::{
+    ffi::c_int,
+    sys_types::pthread_mutex_t,
+};
+
+//==================================================================================================
+// Standalone Functions
+//==================================================================================================
+
+///
+/// # Description
+///
+/// Locks a mutex.
+///
+/// # Parameters
+///
+/// - `mutex`: Mutex object.
+///
+/// # Returns
+///
+/// If successful, zero is returned. Otherwise, an error code is returned instead.
+///
+/// # Safety
+///
+/// This function is unsafe because it may dereference raw pointers.
+///
+/// It is safe to call this function if the following conditions are met:
+///
+/// - `mutex` points to a valid `pthread_mutex_t` structure.
+///
+#[unsafe(no_mangle)]
+pub unsafe extern "C" fn pthread_mutex_lock(mutex: *mut pthread_mutex_t) -> c_int {
+    // Check if `mutex` is not valid.
+    if mutex.is_null() {
+        ::syslog::error!("pthread_mutex_lock(): invalid mutex pointer");
+        return ErrorCode::InvalidArgument.get();
+    }
+
+    match crate::pthread::pthread_mutex_lock(&mut *mutex) {
+        Ok(_) => 0,
+        Err(error) => {
+            ::syslog::error!("pthread_mutex_lock(): failed to lock mutex (error={:?})", error);
+            error.code.get()
+        },
+    }
+}
