+++ src/libs/syscall/src/pthread/bindings/pthread_rwlock_wrlock.rs
@@ -5,21 +5,20 @@
 // Imports
 //==================================================================================================
 
-use ::sysapi::{
-    ffi::c_int,
-    sys_types::pthread_key_t,
+use ::sysapi::ffi::{
+    c_int,
+    c_void,
 };
-use ::syscall::pthread;
 
 //==================================================================================================
-// pthread_key_delete()
+// Standalone Functions
 //==================================================================================================
 
+// TODO: add description
 #[allow(clippy::missing_safety_doc)]
 #[unsafe(no_mangle)]
-pub unsafe extern "C" fn pthread_key_delete(key: pthread_key_t) -> c_int {
-    match pthread::pthread_key_delete(key) {
-        Ok(()) => 0,
-        Err(error) => error.code.get(),
-    }
+pub unsafe extern "C" fn pthread_rwlock_wrlock(_rwlock: *mut c_void) -> c_int {
+    // TODO: https://github.com/nanvix/nanvix/issues/718
+    ::syslog::warn!("pthread_rwlock_wrlock(): not implemented");
+    0
 }
