+++ src/libs/syscall/src/pthread/mod.rs
@@ -29,5 +29,6 @@ cfg_if::cfg_if! {
         pub use syscall::pthread_key_delete;
         pub use syscall::pthread_setspecific;
         pub use syscall::pthread_getspecific;
+        pub mod bindings;
     }
 }
